third_party/WebKit/LayoutTests/typedcssom/inlinestyle/inlineStylePropertyMap_iteration-expected.txt

This is a testharness.js-based test.
PASS Iteration over empty StyleMap returns empty iterator 
FAIL Iterator for single entry returns iterator with a single value assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
PASS Iterator for single key returns iterator with a single value 
FAIL Iterator for single value returns iterator with a single value assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
PASS Iterating entries over border element expansion 
PASS Variable values come out as CSSUnparsedValues 
PASS Custom properties set on the element come out as CSSStyleValues 
PASS @apply rules come out as CSSStyleValues 
Harness: the test ran to completion.
